requirements.txt
amqp==5.1.1
argh==0.28.1
celery==5.3.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-filter==23.2
django-oauth-toolkit==2.3.0
django-pytest==0.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
flower==1.2.0
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-environ==0.4.54
redis==4.5.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
watchdog==2.3.1
sendgrid~=6.10.0
asgiref~=3.7.2